Make separate files for styles, fix responsive design, replace task page with index
