+++ src/components/Layout/Layout.scss
@@ -1,3 +1,5 @@
 .layout {
-  margin: 50px;
+  margin: 0 auto;
+  font-family: roboto;
+  max-width: 100rem;
 }
