+++ src/components/tasks/card/card.scss
@@ -1,0 +1,43 @@
+.card {
+  border: 0.2rem solid black;
+  border-radius: 0.5rem;
+  padding: 1rem;
+  display: flex;
+  flex-direction: column;
+  flex-wrap: wrap;
+  max-width: 25rem;
+  margin-bottom: 2rem;
+  cursor: pointer;
+}
+
+.prTitle {
+  color: #041484;
+  font-size: 1.5rem;
+  font-weight: bold;
+}
+
+.cardFooter {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+}
+
+.profilePicture {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 6rem;
+}
+
+.profilePicture > img {
+  width: 3rem;
+  border-radius: 50%;
+}
+
+.statusLable {
+  color: grey;
+  font-weight: bold;
+}
