+++ src/components/tasks/card/index.tsx
@@ -1,4 +1,5 @@
 import { FunctionComponent } from 'react';
+import classNames from './card.scss'
 
 type pullRequest = {
   title: string,
@@ -24,71 +25,24 @@ const Card: FunctionComponent<Props> = ({ pullRequest }) => {
   } = pullRequest
 
   return (
-    <>
-      <style jsx>{`
-        .card {
-          border: 0.2rem solid black;
-          border-radius: 0.5rem;
-          padding: 1rem;
-          display: flex;
-          flex-direction: column;
-          flex-wrap: wrap;
-          max-width: 25rem;
-          margin-bottom: 2rem;
-          cursor: pointer;
-        }
-
-        .prTitle {
-          color: #041484;
-          font-size: 1.5rem;
-          font-weight: bold;
-        }
-
-        .cardFooter {
-          display: flex;
-          flex-direction: row;
-          flex-wrap: wrap;
-          justify-content: space-between;
-          align-items: center;
-          width: 100%;
-        }
-
-        .profilePicture {
-          display: flex;
-          justify-content: space-between;
-          align-items: center;
-          width: 6rem;
-        }
-
-        .profilePicture > img {
-          width: 3rem;
-          border-radius: 50%;
-        }
-
-        .statusLable {
-          color: grey;
-          font-weight: bold;
-        }
-      `}</style>
-      <div className="card">
-        <span className="prTitle">{title}</span>
-        <span className="statusElement">
-          <span className="statusLable">Estimated completion:</span><strong> {completionDate}</strong>
-        </span>
-        <span className="statusElement">
-          <span className="statusLable">Started:</span><strong> {startedAt}</strong>
-        </span>
-        <div className="cardFooter">
-          <div className="profilePicture">
-            <img src={profilePicture} alt="" />
-            <strong>{author}</strong>
-          </div>
-          <span className="statusElement">
-            <span className="statusLable">Status:</span><strong> {issueStatus}</strong>
-          </span>
+    <div className={classNames.card}>
+      <span className={classNames.prTitle}>{title}</span>
+      <span className={classNames.statusElement}>
+        <span className={classNames.statusLable}>Estimated completion:</span><strong> {completionDate}</strong>
+      </span>
+      <span className={classNames.statusElement}>
+        <span className={classNames.statusLable}>Started:</span><strong> {startedAt}</strong>
+      </span>
+      <div className={classNames.cardFooter}>
+        <div className={classNames.profilePicture}>
+          <img src={profilePicture} alt="" />
+          <strong>{author}</strong>
         </div>
+        <span className={classNames.statusElement}>
+          <span className={classNames.statusLable}>Status:</span><strong> {issueStatus}</strong>
+        </span>
       </div>
-    </>
+    </div>
   )
 }
