+++ src/components/tasks/section/index.tsx
@@ -1,6 +1,6 @@
 import { FunctionComponent } from 'react';
-import Card from '../card'
-
+import Card from '../card';
+import classNames from './section.scss';
 type pullRequest = {
   title: string,
   completionDate: string,
@@ -22,35 +22,12 @@ const Section: FunctionComponent<Props> = ({ heading, content }) => {
   })
 
   return (
-    <>
-      <style jsx>{`
-        .heading {
-          font-size: 3.5rem;
-          font-weight: bold;
-          text-align: center;
-          color: #041484;
-          margin-bottom: 2rem;
-        }
-
-        .section {
-          height: 100%;
-        }
-
-        .cardContainer {
-          display: flex;
-          flex-direction: column;
-          justify-content: space-evenly;
-          align-items: center;
-          flex-grow: 1;
-        }
-      `}</style>
-      <div className="section">
-        <div className="heading">{heading}</div>
-        <div className="cardContainer">
-          {cards}
-        </div>
+    <div className={classNames.section}>
+      <div className={classNames.heading}>{heading}</div>
+      <div className={classNames.cardContainer}>
+        {cards}
       </div>
-    </>
+    </div>
   )
 }
