+++ src/components/tasks/section/section.scss
@@ -1,0 +1,19 @@
+.heading {
+  font-size: 3.5rem;
+  font-weight: bold;
+  text-align: center;
+  color: #041484;
+  margin-bottom: 2rem;
+}
+
+.section {
+  height: 100%;
+  min-width: 17rem;
+}
+
+.cardContainer {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-evenly;
+  align-items: center;
+}
