+++ src/pages/index.tsx
@@ -1,15 +1,31 @@
-import { Component } from 'react';
+import { FunctionComponent } from 'react';
 import Layout from '../components/Layout';
 import Navbar from '../components/navbar';
+import Section from '../components/tasks/section';
+import pullRequests from '../../mock/pullRequests';
+import classNames from '../styles/index.scss';
 
-class Index extends Component<any> {
-  render() {
-    return (
-      <Layout>
-        <Navbar page="Index" />
-      </Layout>
-    );
-  }
+const Index: FunctionComponent = () => {
+  const completedTasks = pullRequests.filter((pr) => {
+    if (pr.completionStatus === 'completed') {
+      return pr
+    }
+  })
+
+  const incompleteTasks = pullRequests.filter((pr) => {
+    if (pr.completionStatus !== 'completed') {
+      return pr
+    }
+  })
+  return (
+    <Layout>
+      <Navbar page="Tasks" />
+      <div className={classNames.container}>
+        <Section heading='Active' content={incompleteTasks} />
+        <Section heading='Completed' content={completedTasks} />
+      </div>
+    </Layout>
+  );
 }
 
 export default Index;
