+++ src/styles/index.scss
@@ -1,0 +1,15 @@
+body {
+  margin: 0 auto;
+}
+
+html {
+  font-size: 15px;
+  margin: 0 5rem;
+}
+.container {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-evenly;
+  align-items: flex-start;
+  flex-wrap: wrap;
+}
